Refactor. Moved exception package from api to upper level. Put all exceptions there
